refactor(hardware/raspb0.nix): remove unnecessary grub and extlinux configurations and update kernel packages to use the default linuxPackages
The previous configuration for the boot loader was overly specific and not required for this hardware. By removing the grub and extlinux configurations, we simplify the boot process. Additionally, updating the kernel packages to use the default `pkgs.linuxPackages` ensures that the system uses the most appropriate kernel for the Raspberry Pi 4, simplifying maintenance and reducing potential conflicts.

fix(hosts/raspb0/configuration.nix): comment out pi5-os-check module

The `pi5-os-check.nix` module is no longer necessary as the system is now configured to run on a Raspberry Pi 4, not a Pi 5. This change prevents potential compatibility issues and ensures the system boots correctly.