refactor(networking): make network interface configurable
This commit refactors the networking configuration to allow for a configurable network interface. Previously, the network interface was hardcoded to "eth0" in several places.

The changes include:
- Introducing a `networkInterface` argument in `modules/networking.nix`, `modules/k3s_agent.nix`, and `modules/k3s_server.nix`.
- Passing this `networkInterface` argument from the host-specific configuration files (`hosts/*/configuration.nix`) to the modules.
- Updating the network interface references within the modules to use the new `networkInterface` variable.

This change improves flexibility and maintainability by allowing different network interfaces to be used without modifying the core module logic. It also addresses the specific case of `raspb0` which uses `end0` as its network interface.