fix(k3s_server_fleet.nix): update fleet helm chart version to 0.13.2 and add fleet-crd chart
This commit updates the version of the fleet helm chart to 0.13.2 to include the necessary CRDs for fleet to function correctly. It also adds the `fleet-crd` chart to the `autoDeployCharts` configuration, ensuring that the CRDs are deployed before the main fleet chart. This resolves an issue where fleet was not functioning due to missing CRDs.