perf(k3s_server_fleet.nix): increase polling interval for gitops to 120s
This change increases the polling interval for the gitops configuration from 30 seconds to 120 seconds. This is done to reduce the frequency of git operations, which can be resource-intensive and potentially cause unnecessary load on the cluster. By increasing the interval, we aim to improve the overall performance and stability of the K3s server fleet.