refactor(modules): move metallb manifests to k3s_server_metallb.nix
The MetalLB setup was previously handled by a systemd service in the host configuration. This commit moves the MetalLB manifests to the `k3s_server_metallb.nix` module. This change centralizes the MetalLB configuration within the Kubernetes-specific module, making it more organized and easier to manage. The systemd service in the host configuration is removed as it is no longer needed.