fix(configuration.nix): use unstable channel for libnvidia-container to get newer version
This change updates the `libnvidia-container` package to use the unstable Nixpkgs channel. This is necessary to obtain a newer version of `libnvidia-container`, which is required for compatibility with certain Kubernetes or Docker workloads. By overriding the default package with one from the unstable channel, we ensure that the server has access to the latest features and bug fixes for NVIDIA container management.