refactor(flake.nix, hosts/server/configuration.nix): use nixpkgs unstable input for newer packages
This change moves the sourcing of the `nixos-unstable` channel from within the host configuration to the flake inputs. This allows for easier management and access to newer packages across different hosts and configurations. The `unstable` input is now passed as a special argument to the host configuration, simplifying the import process and improving maintainability.